fix(SR-1): mock-data-loader uses endpoint instead of directly updating DB.
