+++ utils/mock-data-loader/.env.sample
@@ -2,7 +2,4 @@ DEAL_API_URL=http://localhost:5001
 DTFS_CENTRAL_API=http://localhost:5005
 TFM_API=http://localhost:5004
 
-MONGO_INITDB_DATABASE=
-MONGODB_URI=
-
 API_KEY=test
