+++ utils/mock-data-loader/package.json
@@ -21,7 +21,6 @@
     "dotenv": "^16.3.1",
     "generate-password": "^1.7.0",
     "json2csv": "^5.0.7",
-    "moment": "^2.29.4",
-    "mongodb": "4.2.0"
+    "moment": "^2.29.4"
   }
 }
